fixed a bunch of shit
updating arrows work as long as you're not spam killing the monsters,
which isn't good. My guess is that update is called so many times when
you're clicking the monsters that it crashes since if you tap slowly it
loads fine, but there is a slight delay.

Fixed the upgrade issue where the counters were starting from 0, even
tho i had set them to start from 1. The problemet existed in the
loadState where the default value was set to 0.
"git blame Erik Sjöström"

Fixed a little ui problem where the text seemed to be limited to an area
so that whenever a value exceeded 100 it was new line the remaining
value. hopefully from this fix, the textview field will extend instead
of newlining